Added the concept of asset producers, addressing #17 Refactored module handling into a producer. Refactored unit tests.
