+++ src/main/java/org/terasology/assets/module/AssetDeltaFormat.java
@@ -14,11 +14,12 @@
  * limitations under the License.
  */
 
-package org.terasology.assets;
+package org.terasology.assets.module;
 
+import org.terasology.assets.AssetData;
+import org.terasology.assets.AssetInput;
 import org.terasology.assets.exceptions.InvalidAssetFilenameException;
 import org.terasology.naming.Name;
-import org.terasology.naming.ResourceUrn;
 
 import java.io.IOException;
 import java.nio.file.PathMatcher;
@@ -27,7 +28,7 @@ import java.util.List;
 /**
  * @author Immortius
  */
-public interface AssetFormat<T extends AssetData> {
+public interface AssetDeltaFormat<T extends AssetData> {
 
     /**
      * @return A path matcher that will filter for files relevant for this format.
@@ -35,19 +36,18 @@ public interface AssetFormat<T extends AssetData> {
     PathMatcher getFileMatcher();
 
     /**
-     * @param filename The filename of an asset, including extension
+     * @param filename The filename of an asset delta file
      * @return The asset name corresponding to the given filename
      * @throws InvalidAssetFilenameException if the filename is not valid for this format.
      */
     Name getAssetName(String filename) throws InvalidAssetFilenameException;
 
     /**
-     * Loads an asset.
+     * Applies a delta to the given assetData
      *
-     * @param urn    The urn identifying the asset being loaded.
-     * @param inputs The inputs corresponding to this asset
-     * @return The loaded asset
-     * @throws IOException If there are any errors loading the asset
+     * @param inputs    The inputs corresponding to this asset
+     * @param assetData An assetData to update
+     * @throws java.io.IOException If there are any errors loading the delta
      */
-    T load(ResourceUrn urn, List<AssetInput> inputs) throws IOException;
+    void applyDelta(List<AssetInput> inputs, T assetData) throws IOException;
 }
